dpkg: Always spawn a new shell on conffile prompt
Stop supporting self backgrounding, remove DPKG_NO_TSTP environment
variable support, as we always spawn a shell now.

This will allow setting useful environment variables for the user to use
with other tools. And is less confusing, as the action to go back to
dpkg does not depend on how dpkg was run.

Closes: #38334