Remove --license and --licence options from tools
The options are pretty useless and non-standard. The user can surely
cat a file from the installed package. In addition this option has not
worked ever on non-Debian systems as it's relying on the GPL-2 file
from the base-files Debian package. So remove it, and we get some code
reduction as a nice side-effect.