Dpkg::Source::Patch: add more sanity checks on patches
patch will happily accept filenames like "../../../../stuff" and modify
files outside of the expected destination directory. To avoid problems
we error out when we detect a filename that contains "/../". Any leading
"../" is not a problem since patches are applied with -p1 and it's
stripped.

We also verify that the file to be modified is not accessed through a
symlink as a compromised source package could also provide a symlink in
the orig.tar.gz that points outside of the destination directory.