libdpkg: Encapsulate triggersdir handling in triglib
The only users of triggersdir and related files are the triglib and
trigdeferred modules, and dbmodify does not have any business in knowing
where the triggersdir is located (besides for conveninence when creating
the pathname). Create instead a new function trig_get_triggersdir() to
generate the triggersdir from an admindir, and use that on each module.