dselect: treat unknown packages as known and marked for purge
This is a temporary work-around so that dselect doesn't try to
reinstall packages of priority > standard that were removed or not
installed.

The complete solution (tracked in #551638) will involve storing the
seen/not-seen information somewhere else than in the dpkg status database
and ideally in a new database shared by all frontends.