Fix installation of disappearing replaced packages in reverse order
This happens when the we install first the replacing then the replaced
package, for which the replaced package is supposed to get disappeared.
And fixes it to disappear the correct package and not lose track of the
ownership of the replaced files, by marking the replaced file as not
being part of the unpacked archive.