Dpkg::Version: fix boolean evaluation of the object
In a boolean context, the object evaluates like the version string if
the version is valid, otherwise it evaluates like undef. Document
this behaviour and add non-regression test to ensure it's kept.

This implicitly fix dpkg-shlibdeps who relies on the boolean evaluation
of the object to decide whether or not a minimal version specification
is needed or not.