dpkg-shlibdeps(1): avoid triggering a lintian warning on translations
Translations of the manual pages end up with "\-" instead of "-" that
is in the original manual page and as such man is not allowed to break
the long path name there.

Avoid all the problems by using a shorter path in the example so that the
path doesn't end up alone on its line without any possibility to adjust on
the right.

Update the translations at the same time to avoid some fuzzy strings.