dpkg-maintscript-helper: add "supports" command
With this command a maintainer script can verify whether a given command
is supported by dpkg-maintscript-helper before calling it. Thanks to this
he can avoid a pre-dependency on dpkg.

Also improve the error message output when the command is unknown to
suggest upgrading dpkg as a possible way to fix the problem encountered.