dpkg: Add two new dpkg options --path-exclude and --path-include
This provides support for filtering paths on package installation. This
allows embedded systems to skip /usr/share/doc, manpages, etc.

dpkg does not lose track of excluded paths during filtering, and they
get checked for file conflicts as usual, so filters are not a way to
avoid file conflict situations.

Closes: #68788, #68861, #497304, #525567, #583902

Based-on-patch-by: Tollef Fog Heen <tfheen@err.no>
Signed-off-by: Martin Pitt <martin.pitt@ubuntu.com>
Signed-off-by: Guillem Jover <guillem@debian.org>