Don't hardcode the list of tests
We use $(wildcard t-*) to discover all tests. Drop the only useless file
that was matched by the wildcard and that is not a real test case.

Several tests were missing from the list and with this change this should
not happen any more.

The order of the tests should not matter and if a specific order
is important (say testing basic operations first), then the directories
should be renamed to include a numbered prefix (t-XXXX-short-desc).