libdpkg: update w_booleandefno() to cope with the changed type of booleans
Boolean fields are now stored in "bool" variables and no longer in integers.
The former is 1-byte long and doesn't match the length of an int, so the
cast done in PKGPFIELD was reading too much data after the offset where
the boolean field is stored.

This update was missed in commit 7eb30624a0b7955924bafd9466d226f70e5cf48f.