build: Use DPKG_CHECK_DECL instead of DPKG_CHECK_DEFINE for TIOCNOTTY
DPKG_CHECK_DEFINE defines the HAVE_symbol to 1 if found, and to 0
otherwise. But the only current user is doing “#ifdef TIOCNOTTY”, which
defeats the test as it's always going to be true, regardless of the
value. Switch to the simplified DPKG_CHECK_DECL, which does not define
the symbol if not found, and remove now unused DPKG_CHECK_DEFINE.