build: Do not fallback to use chown instead of lchown if not available
When creating symlinks on extraction, dpkg needs to set its owner, if
we fallback to use chown in place of lchown we might get to a situation
where the dereferenced symlink does not yet exist, which would cause an
unwanted error.

On some older and buggy systems, chown does not actually dereference
symlinks, so it could be safely used although it would need to be
detected at runtime, as the kernel could be switched under our feet,
but given that lchown is already specified by UNIX98, there's not much
sense in not assuming it's present and working at this point in time.