dpkg-source: new --unapply-patches option
This option is used by source formats 2.0 and 3.0 (quilt) to unapply
patches after a successful build. It's useful for people who manage their
packages in a VCS repository and prefer to have patches unapplied, even
after a package build.

Given that this option is a matter of personal taste, do not allow it
in debian/source/options so that the resulting source package cannot
have this behaviour by default and so that all source packages behave
in a consistent way: patches applied by default is a major feature
of the new formats.