dpkg: Remove filtered files on upgrade
Because the filtered file is left in the new file list, the code that
verifies if the old file is present (maybe with a different name) in
the new list matches the stat information. So we mark and treat filtered
files as if they were already not present on the file system.

Reported-by: Martin Pitt <martin.pitt@ubuntu.com>
Signed-off-by: Guillem Jover <guillem@debian.org>