Add new --force-unsafe-io to disable safe I/O operations on unpack
This allows to not perform file system syncs before file renames
to guarantee its atomicity, which is known to cause substantial
performance degradation on some file systems, unfortunately the ones
that require the safe I/O on the first place due to their unreliable
behaviour causing zero-length files on abrupt system crashes (sudden
reboot, bus locks, pulling the plug, etc).

Using this option might improve performance at the cost of losing
data, and should thus be used with care, but that's ultimately
something for the user of the affected file systems to decide.

Closes: #584254