build: Disable usage of synchronous sync(2) by default
It causes undesired I/O on unrelated file systems. It also makes the
code behave differently on Linux systems.

Allow the possibility to enable it again for the benefit of downstreams,
which might want to use it regardless of its problems. Although this
code path will most probably be removed in the near future.

Closes: #588339, #595927, #600075