libdpkg: When parsing database files only warn on bogus versions
Those versions, although bogus, were previously accepted, so to avoid
breaking systems by making dpkg refuse to parse the status and available
files, we just make it warn for now, and will make it error out on
status file parsing later on.

We keep producing errors on the rest of version parsing to avoid newly
introduced bogosity and so that packages get fixed, while not making the
system unusable anymore.

Closes: #590885, #590896, #591692, #591885