Fix install target in man/Makefile.am
Commit 39c6dab89bbea9fe336f869b65e33102ba238205 introduced a regression:
make install during a package build in a tree generated by make dist would
no longer install the manual page... because they are already built and
available in $(srcdir) while $(CURDIR) was ok for the case where the
manual pages are not pre-built.

No we try both paths and pick the first one that exists.