Dpkg::Shlibs::Objdump::get_format() fallback to host objdump if cross one failed
In many cases the cross objdump is not able to analyze the binaries from
the build host and yet dpkg-shlibdeps need to be able to verify that
the various binaries use the same ELF format. To achieve this get_format()
is enhanced to fallback on the objdump command when the cross one failed.