dpkg-genchanges: correctly truncate descriptions with multibyte characters
Ensure the scalar used to truncate the description is character-based
and not byte-based. But switch it back to a byte-based scalar afterwards
to avoid bad-conversion to latin1 when output in a filehandle without
any explicit encodind.

This should really be fixed in Dpkg::Control but that would be an invasive
change at this point of the squeeze release.

Reported-by: Colin Watson <cjwatson@ubuntu.com>