dpkg-source: drop -k parameter from the tar command line used to extract tarballs
dpkg-source only extracts tarballs in directories that it has created
itself, there's no risk of overwriting any user files. Furthermore in
the few cases where we do extract a tarball on a non-empty directory,
we really want to be able to overwrite files already present. That's the
case with a debian.tar.gz file.

Reported-by: James Westby <james.westby@linaro.org>