libdpkg: Add error context function handler support
This allows the code to use a function instead of a jump buffer to
handle the errors, which happen to be possible for the general case,
when it does not need to continue processing in an upper stack frame.
Thus reducing the preceived complexity of simple code, and avoids
exposing setjmp machinery to the casual user.