libdpkg: Fix version_strict_check() to not allow '-' in revision
The revision cannot possibly have a '-' in it, so do not allow it. This
is innocuous in the current code as the parser makes sure the versions
are split on last '-', which implies the revision will never have one,
but it could happen that a manually constructed versionrevision could
contain one.