dpkg: Defer symlink renames to avoid ending up with broken symlinks
The broken symlinks are particularly pernicious as they imply a shared
library SONAME path might point to nowhere and thus applications will
fail to get dynamically linked.

Deferring symlinks is safe as they are always at the end of the tar
archive, and they are processed last on deferred renames as they are
appended at the end of the new list files.

Closes: #605536