Use m_asprintf() instead of nfmalloc() and string operations
This avoid the need to know the resulting string length beforehand, and
makes sure it's always going to get the right size. It also makes the
code way clearer.

Usage of nfmalloc() here was not appropriate anyway, as the variables
are only of temporary use, so switching to normal allocation gives
a side-effect “bug” fix.