libdpkg: Implement a centralized list of architectures
The goal is to avoid having hundreds of strdup'ed architecture strings in
memory when we can instead store some pointers to a new structure. Those
structs are singletons and it's thus now possible to compare the pointers
directly if we want to compare two architectures. Furthermore the struct
allows us to store conveniently some basic information like if the
architecture is the native one, a foreign one, a special one (all/any),
an illegal one, no architecture present or something completely unknown.

Sponsored-by: Linaro Limited

[guillem@debian.org:
 - Add an explicit test for dpkg_arch_reset(), and do not rely on
   pkg_db_reset().
 - Check that the default list has exactly 3 items instead of >= 3.
 - Change call to dpkg_arch_reset() before nffreeall().
 - Rename dpkg_arch_reset() to dpkg_arch_reset_list(). ]

Signed-off-by: Guillem Jover <guillem@debian.org>