libdpkg: Use a varbuf for the field_state value
This should reduce the amount of m_realloc() calls done when parsing,
as the varbuf keeps track of the memory allocated and does not alternate
reducing and growing the buffer unnecessarily.

It might allow in the future to avoid copying around the field value
string just to be able to append a NUL character.