Propagate --admindir to programs run from maintainer scripts
Always set DPKG_ADMINDIR from dpkg so that all subprocesses inherit
the variable and there's no possible security risk from a user process
setting the variable to write onto a user controlled area.

As usual, the order in which the admindir value is overridden is the
following: built-in default, environment variable and command line
option.

Closes: #97076

Suggested-by: Brendan O'Dea <bod@debian.org>