build: Add support for DPKG_TESTSUITE_OPTIONS local-db
This makes the test-suite run extremely faster than using the build
system. It will also be less prone to getting the build system database
into inconsistent state. And lastly, it will not interfere and thus not
lock the build system database.

Not enabled by default for now, because this requires the fix for dpkg
to propagate the --admindir to its childs, and the fix to not fail when
removing the root directory, which are still in the git master branch.