debian/control: drop Breaks on konqueror
With this Breaks, "apt-get install udev" that is recommended at the start
of a lenny->squeeze upgrade might remove the kde metapackage. This is
due to udev needing a new util-linux that itself needs a newer dpkg (due
to the install-info transition).

In contrast, the lack of the Breaks has very minimal consequences: partial
upgrades might not install install-info and in that case konqueror might
show an outdated list of info pages.