Update deppossi->ed to point to a pkgset instead of a pkginfo
It represents only the package name of the dependency. Later deppossi
will grow an architecture parameter that will allow us to transform
that pkgset into the correct pkginfo (depending on the context).

Note that this transformation depends on the dependency type
(possi->up->dep_type) and as such it won't be done at parsing time
but rather every time that we need to analyze a specific dependency.

[hertzog@debian.org: (Sponsored-by: Linaro Limited)
 - Update code for new location of depended struct.
 - Update call sites for pkg_db_find() returning pkginfo. ]

Signed-off-by: Guillem Jover <guillem@debian.org>