libdpkg: Add new multiarch aware in-core database functions
The new functions are:

- pkg_db_find_set(name)
- pkg_db_find_pkg(name, arch)
- pkg_db_iter_next_set(i)
- pkg_db_iter_next_pkg(i)
- pkg_db_count_set()
- pkg_db_count_pkg()

And rework the legacy functions pkg_db_count(), pkg_db_find(name) and
pkg_db_iter_next(name) to work on top of the new functions.

The pkg_db_find_pkg() function is just a compatibility stub for now,
so that there's no point in time when dpkg can load multiple package
instances but not save them to disk, which would imply losing data.

Sponsored-by: Linaro Limited

Signed-off-by: Guillem Jover <guillem@debian.org>