libdpkg: Ensure parsed packages have a usable arch member
In case the package didn't have an Architecture field, warn and
initialize it to an empty string, so that it can be safely used for
string comparisons, printing, etc.

This fixes a potential segfault when setting the environment variable
DPKG_MAINTSCRIPT_ARCH from a package w/o an Architecture field, which
are a thing from the distant past, but still should be supported.