libcompat: Define gettext context functions when NLS is disabled
When the package is configured with NLS disabled (--disable-nls),
the build will fail due to undefined references to LC_MESSAGES, the
gettext.h header file does not correctly handle the gettext context
related function declarations on disabled NLS support, which is known
to be used by downstreams. This is the minimal change needed to fix
the build failure. A more complete fix will be deployed on the master
branch.