dpkg: Complete arch checking of package relationships in the resolver
In places where the architecture qualifier of a dependency has not been
validated before-hand (through usage of deppossi_pkg_iterator), add a
manual verification.

In both cases, the checks are not needed with the current rules where
both Replaces and Breaks can only have “any” as architecture qualifier
but for completeness, we put them nevertheless.

Signed-off-by: Guillem Jover <guillem@debian.org>