dpkg: Automate debug help output printing
Add a new array of struct debuginfo to hold debug flag information, so
that it can be handled automatically.

This has multiple purposes; it will guarantee the output value always
matches the one defined in the code, there's one less place to modify
when adding a new debug flag, the ouput is easier to modify w/o fuzzying
the translated strings, it's easier for the translators to deal with new
debug flag additions, it will not allow for the source reference name to
be translated thus avoiding errors.