dpkg: Automate force help output printing
Move the force type (normal, default, dangerous) and force description
to struct forceinfo, so that its printing can be automated.

This has multiple purposes (similar to the debuginfo output automation);
there's one less place to modify when adding a new force flag, the ouput
is easier to modify w/o fuzzying the translated strings, it's easier for
the translators to deal with new force flag additions.