dpkg-source: use server-side clock for patched file timestamp on NFS
Since 1.13.14~20 (2006-02-10), dpkg-source touches the files it
patches when unpacking, with a single date.  This way, the order of
mtimes does not depend on the order in which the files were patched,
which is convenient when e.g. configure.in and configure are patched.

More precisely, dpkg-source does something like the following:

	my $now = time();
	foreach my $fn (@patched_files) {
		utime($now, $now, $fn);
	}

Unfortunately when the filesystem is NFS, "touch" and normal
modification set mtime and atime to the current time on the server
side, while time() returns the current time on the client side.
The two clocks can disagree, producing breakage.

So unless a timestamp has been passed explicitly, use utime(undef,
undef, $fn) to set mtime for the first file to the server side time
and copy it (rounded down to a number of seconds) to all patched
files.

Reported-by: Stéphane Glondu <glondu@debian.org>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Improved-by: Raphaël Hertzog <hertzog@debian.org>
Signed-off-by: Raphaël Hertzog <hertzog@debian.org>