dpkg-divert: Properly handle diversions of shared files
dpkg-divert does nothing if the same diversion is recorded multiple
times. However when a diversion is removed, it does so even if
the shared file is still in place.

We change dpkg-divert's behaviour to ignore a diversion removal request
until the diverted file is no longer owned by another package of the
same package set.

We also update the test-suite to provide a status file now that
dpkg-divert reads it.

Sponsored-by: Linaro Limited

[guillem@debian.org:
 - Handle missing DPKG_MAINTSCRIPT_ARCH when upgrading from old dpkg.
 - Split ignore removal logic into a new function. ]

Signed-off-by: Guillem Jover <guillem@debian.org>