dpkg: Add multiarch to monoarch database layout downgrade support
Handle the case where downgrading would lose data, i.e. when there's
co-installed packages which the old dpkg would not be able to handle,
and abort then.

If the downgrade is safe, just convert the multiarch infodb to a
monoarch infodb layout in two stages, first create hardlinks from the
arch annotated control files before the upgrade starts, and then once
the new package is in place, remove the arch annotated control file
leftovers. This way if the upgrade got interrupted we always have a
working database for both versions of dpkg.

[hertzog@debian.org:
 - Detect arch qualified dependencies in status and available files. ]