Do not read the available file unnecessarily
For commands that only make use of information from the status file,
there's no need to make them read and parse the available file. This
change does not alter the current behaviour, and commands that were
previously updating the available file will continue doing so.

Removing further access to it is pending a decision on how to treat
globally the available file, which needs interaction with the package
manager frontend developers.

Closes: #397121

Based-on-patch-by: Michel Lespinasse <walken@zoy.org>