dpkg-source: extend the Package-List field based on the feedback received
Integrate the architecture field with spaces replaced by commas. It will
be used by the wanna-build team to see if the source package builds
Architecture: all packages. The reprepro maintainer also had a similar
use case for this information.

Put the type of the package in a dedicated field rather than a prefix of
the package name. Suggested by Bastian Blank.

See feedback at
http://lists.debian.org/debian-policy/2011/03/threads.html#00155