dpkg-split: Do not corrupt binary part metadata on 32-bit systems
On 32-bit systems the format string used to generate the metadata
archive member of a split package specifies a 32-bit value to be
retrieved from va_arg, but the variable passed is a 64-bit one, which
messes with the layout of the stack, producing garbage on output.

Regression introduced in commit f2115151c19ff37b305296c23225807b0832086b.