libdpkg: Include <locale.h> from <i18n.h> to fix unoptimized builds
The pgettext() macros in gettext.h make use of LC_MESSAGES, which get
normally pulled from <locale.h> through <libintl.h>, but that only
happens on optimized builds. As this is an implementation detail,
pgettext() users (through the local C_() macro) should not need to
know they have to include <locale.h>. Arguably this is a deficiency
in gettext.h which we are working around here.

Reported-by: Steve Langasek <steve.langasek@ubuntu.com>