update-alternatives: new syserr() function
This function is to be used when a syscall failed and when we want
to print strerror(errno) after the error message.

Update all the code to use this function. This allows to reduce the
number of strings to translate as we're now reusing the same strings
than those passed to oshite() in dpkg.