libdpkg: Allow again unknown Priority field values
The code was not properly advancing the end of the value string when
reaching the fallback case (on values not known for the Priority field),
and the validation for junk after the first word was failing on the
actual first word.

This change reverts to the previous behaviour of blindly accepting the
whole value string for a field, regardless of it being multiword, when
it's the fallback case.

Regression introduced in commit b3f669039f128d715ac7ac71abaeac86f0954112.

Analysis-by: Raphaël Hertzog <hertzog@debian.org>
Analysis-by: Guillem Jover <guillem@debian.org>