dpkg: Fix wrong matching for check on directory containing conffiles
On package removal, keep only directories actually containing conffiles,
and not directories just matching the substring in the conffile or the
directory itself. The directory itself is not matched either as strictly
speaking it is not contained in itself, and it should already be in the
leftover lists.

Previously when checking for directory ‘/foo’ the following conffiles
would match:

  /foo
  /foobar
  /foo/bar

Now only the following will match:

  /foo/bar

Spotted-by: Ondřej Surý <ondrej@debian.org>