dpkg-source: fix behaviour of --extend-diff-ignore if used after --diff-ignore
Up to now --extend-diff-ignore was only extending the default diff-ignore.
This is usually enough as this option is usually put in
debian/source/options and interpreted early enough before any other -i
option (that activates the usage of the default diff-ignore regexp).

Should --extend-diff-ignore appear after -i, it should still extend
the regexp that we're going to use.