dpkg: implement "interest-noawait" and "activate-noawait" trigger commands
Those variants do not put triggering packages in triggers-awaited status
and thus do not record the package with the corresponding pending triggers
in the Triggers-Awaited field.

This should be used for triggers which do not provide essential
functionality when we can safely consider that the triggering packages
are able to satisfy dependencies even if the trigger processing
has not yet happened.