dpkg: Do not preserve and check for conffiles on leftover dir removal
The function removal_bulk_remove_leftover_dirs() was supposed to be
called only w/ left over dirs, but removal_bulk_remove_configfiles()
does not remove them from the file list when unlinking.

Do not check if the directory needs to be preserved if the package has
conffiles inside, because it should have none, although it would match
given that the file list still contains conffiles.

Do not preserve any lingering conffiles in the file list as they are
no longer on disk anyway.