dpkg-source: reenable the Package-List field
But drop the Architecture column since we have no clear use case yet. It
can always be added later on. Parsers should treat the field as an
extendable one. They shall ignore supplementary columns that they do
not know.

Also drop the source line, it's not needed since the dsc file describes
the source package already (source section and priority are not currently
exported in dedicated fields but they can be added later if we start
having a need for this information).