dpkg: ensure that trig_transitional_activate() doesn't introduce bad data
trig_transitional_activate() is emptying Triggers-Pending without
ensuring that the status is set back to something which is not
triggers-pending. In theory it's ok because the trig_parse_ci()
call is supposed to reactivate all the triggers. But if there's no
triggers file any more then we're effectively introducing a skew
which is going to be fatal the next time we try to write down
the status file.

The disparition of the triggers file can happen with file system problems,
bad user manipulation or even an old dpkg bug that kept triggers alive
when in fact the trigger was dropped from the package. Or the triggers
file has effectively been removed in the package but the user has restored
an old status file with a triggers-pending status that the current package
should no longer be able to generate.

In other words, properly resetting the status in
trig_transitional_activate() makes dpkg more robust.