dpkg: do not fallback to "new-prerm failed-upgrade" for downgrades
If "installed-prerm upgrade" fails, dpkg tries to run "new-prerm
failed-upgrade" so that newer versions of packages can work-around
a bug in the prerm of the installed package.

In the case of downgrade this logic doesn't make any sense since
the oldest version can't be aware of the bug in the newest version
(that did not exist when the prerm of the oldest version was
written).

Thus we're disabling the fallback in case of a downgrade or a
reinstallation of the same version.