dpkg-source: add some supplementary checks of debian/control
We're now ensuring that debian/control contains a source entry
and at least one entry for a binary package. And we also ensure
that the Architecture field is present. This a new restriction
but it's really in-line with the change made in dpkg where we
complain of the lack of this field.

Reported-by: Niels Thykier <niels@thykier.net>