dpkg: Unpack missing shared directories when the package is replaced
When unpacking a directory missing in the file system with that path
shared by at least one other package, from a package that is Replaced
by other packages, assume the missing path is a shared directory and
avoid setting ‘keepexisting’ and as such returning prematurely from
tarobject(), because in that case we need to recreate the lost
directory. As a side effect this also prevents printing the message
“Replaced by files in installed package” which is bogus in this case,
as directories are always shared.

Regression introduced in commit 00e5640a99be03aba40c9e08a663b90d8f8aa797.

Closes: #631808