dpkg-buildflags: support debian/buildflags
This file is for use by package maintainers. Since the call to
dpkg-buildflags might be hidden by packaging helpers, it's a good
idea to offer a simple way for package maintainers to extend
the set of flags returned. Otherwise each package helper must provide
its own way to extend the compilation flags.