dpkg-buildflags: new --dump action, make it the default
The --export command is useful to retrieve the compilation flags
in a shell script or in a makefile, and to export them in the environment,
but it's not well suited for parsing by another script (or even by a
human).

The new --dump option becomes the default action and prints
the flags and their values following a documented template: "flag=value".